fix: address PR feedback, merge conflicts, and CI issues
- Refactored `memory-storage-turso/src/storage/mod.rs` to extract internal embedding methods into `embeddings_internal.rs`, maintaining the 500-line limit.
- Switched to async `tokio::fs::create_dir_all` for cache directory initialization in `storage.rs`.
- Updated OAuth security tests to use explicit `Algorithm::HS256` in JWT headers.
- Resolved merge conflicts with the latest main branch.
- Fixed Rust type inference issues (`libsql::Value`) and formatting violations identified by CI.
