+++ memory-mcp/src/bin/server_impl/storage.rs
@@ -60,7 +60,8 @@ pub async fn initialize_redb_only_storage() -> anyhow::Result<Arc<SelfLearningMe
 
     // Create data directory if it doesn't exist
     if let Some(parent) = cache_path.parent() {
-        std::fs::create_dir_all(parent)
+        tokio::fs::create_dir_all(parent)
+            .await
             .map_err(|e| Error::Storage(format!("Failed to create data directory: {}", e)))?;
     }
 
@@ -120,6 +121,13 @@ pub async fn initialize_dual_storage() -> anyhow::Result<Arc<SelfLearningMemory>
     let cache_path_str =
         std::env::var("REDB_CACHE_PATH").unwrap_or_else(|_| "./data/cache.redb".to_string());
     let cache_path = Path::new(&cache_path_str);
+
+    // Create data directory if it doesn't exist
+    if let Some(parent) = cache_path.parent() {
+        tokio::fs::create_dir_all(parent)
+            .await
+            .map_err(|e| Error::Storage(format!("Failed to create cache directory: {}", e)))?;
+    }
 
     let cache_config = CacheConfig {
         max_size: std::env::var("REDB_MAX_CACHE_SIZE")
@@ -182,6 +190,13 @@ pub async fn initialize_turso_local() -> anyhow::Result<Arc<SelfLearningMemory>>
     let cache_path_str =
         std::env::var("REDB_CACHE_PATH").unwrap_or_else(|_| "./data/cache.redb".to_string());
     let cache_path = Path::new(&cache_path_str);
+
+    // Create data directory if it doesn't exist
+    if let Some(parent) = cache_path.parent() {
+        tokio::fs::create_dir_all(parent)
+            .await
+            .map_err(|e| Error::Storage(format!("Failed to create cache directory: {}", e)))?;
+    }
 
     let cache_config = CacheConfig {
         max_size: std::env::var("REDB_MAX_CACHE_SIZE")
