+++ memory-mcp/tests/oauth_security_tests.rs
@@ -38,7 +38,7 @@ fn test_validate_bearer_token_security() {
         scope: Some("mcp:read".to_string()),
     };
     let valid_token = encode(
-        &Header::default(),
+        &Header::new(Algorithm::HS256),
         &my_claims,
         &EncodingKey::from_secret(secret.as_bytes()),
     )
@@ -51,7 +51,7 @@ fn test_validate_bearer_token_security() {
 
     // 2. Test Forged Token (wrong secret)
     let forged_token = encode(
-        &Header::default(),
+        &Header::new(Algorithm::HS256),
         &my_claims,
         &EncodingKey::from_secret("wrong-secret".as_bytes()),
     )
@@ -68,7 +68,7 @@ fn test_validate_bearer_token_security() {
         ..my_claims.clone()
     };
     let expired_token = encode(
-        &Header::default(),
+        &Header::new(Algorithm::HS256),
         &expired_claims,
         &EncodingKey::from_secret(secret.as_bytes()),
     )
@@ -85,7 +85,7 @@ fn test_validate_bearer_token_security() {
         ..my_claims.clone()
     };
     let wrong_iss_token = encode(
-        &Header::default(),
+        &Header::new(Algorithm::HS256),
         &wrong_iss_claims,
         &EncodingKey::from_secret(secret.as_bytes()),
     )
@@ -102,7 +102,7 @@ fn test_validate_bearer_token_security() {
         ..my_claims.clone()
     };
     let wrong_aud_token = encode(
-        &Header::default(),
+        &Header::new(Algorithm::HS256),
         &wrong_aud_claims,
         &EncodingKey::from_secret(secret.as_bytes()),
     )
@@ -134,7 +134,7 @@ fn test_validate_bearer_token_insecure_fallback() {
 
     // Create a token with ANY secret (or none)
     let token = encode(
-        &Header::default(),
+        &Header::new(Algorithm::HS256),
         &my_claims,
         &EncodingKey::from_secret("anyone-can-sign".as_bytes()),
     )
